SBS coupled equations with noise
A solver for the SBS (stimulated Brillouin scattering) coupled mode equations with thermal noise and laser phase noise. This simulation is based on a numerical paper from JOSA B entitled "Numerical Simulation of Noise in Pulsed Brillouin Scattering", details for the paper are found in the comments of the script.